monad_service/requirements.txt
web3
requests
py-solc-x
flask
flask-cors
